solpic_api_server/requirements.txt
aiobotocore==2.17.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aioitertools==0.12.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
black==25.1.0
blessed==1.20.0
cement==2.10.14
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distlib==0.3.9
dnspython==2.7.0
docutils==0.19
email_validator==2.2.0
fastapi==0.115.6
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
gunicorn==21.2.0
h11==0.14.0
huggingface-hub==0.29.1
identify==2.6.8
idna==3.10
image==1.5.33
jmespath==1.0.1
motor==3.1.2
multidict==6.1.0
mypy-extensions==1.0.0
newrelic==9.7.1
nodeenv==1.9.1
packaging==24.2
pathspec==0.10.1
pillow==11.2.1
platformdirs==4.3.6
pre_commit==4.1.0
propcache==0.3.0
pyasn1==0.6.1
pydantic==2.10.4
pydantic_core==2.27.2
pymongo==4.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rsa==4.7.2
semantic-version==2.10.0
setuptools==76.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.41.3
termcolor==2.5.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.28.0
virtualenv==20.29.2
wcwidth==0.2.13
wrapt==1.17.2
yarl==1.18.3
redis==4.6.0
